Add package for dedenting text
It often occurs that we need to write multi-line strings in our code, and
we want to keep the code clean and readable. While Go supports using
backticks for raw multi-line strings, it makes the formatting of the code
look ugly. This package provides a simple function to dedent a multi-line
string, removing leading whitespace from each line.